gesturio/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-redis==5.4.0
django-rest==0.8.7
djangorestframework==3.15.2
dotenv==0.9.9
google-auth==2.38.0
idna==3.10
Markdown==3.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
